Adjust Conda Feedstock and Enable macos arm64 builds
This PR changes a large amount with how the conda feedstock is being built.
It switches python to use setup.py so it can find the libtiledbsoma as build
from the multi-package build. It also adjusts to have build numbers and lastly
enable MacOS arm64 builds. Python 3.7 was also disabled for MacOS due to issues
at runtime with installing the dependency. Downstream users for python 3.7 are
currently limited to linux, so this is an acceptable mitigation.